Use lists more when translating from aliencode to lambda calculus.
This was just a bug. Previously, the code to convert `cons x nil` to the
list `[x]` was used only when `ap ap cons x nil` appeared literally in
the source code, not when the `cons` and the arguments were brought
together by β-reduction.

This affects functions like this one:

    :1112 = ap ap b ap b ap cons 3 ap ap c ap ap b b cons ap ap c cons nil

which was previously rendered in lambda calculus as:

          = (\x -> (\y -> cons 3 (cons x (cons y nil))))

With the change, it looks like this:

          = (\x -> (\y -> [3, x, y]))